feat: add many-to-many relationship between User and Novel,

+++ backend/src/main/java/com/mediafetch/backend/novel/repository/NovelRepository.java
@@ -4,6 +4,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.mediafetch.backend.novel.model.Novel;
 
-public interface NovelRepository extends JpaRepository<Novel,Integer>{
+public interface NovelRepository extends JpaRepository<Novel, Long>{
 
 }
